feat(views): kebab menu on sidebar views, sharing view actions via useView
- Move the per-view action menu (Duplicate/Pin/Show-Hide/Make Public-Private/
  Edit/Delete) and the create/edit/duplicate submit handler into the useView
  composable so the breadcrumb dropdown and the sidebar share one source of
  truth. Each consumer passes its own dialog and renders its own ViewModal.
- AppSidebar: reveal a "..." kebab on hover over a view row (and keep it while
  its menu is open); reserve trailing space so long labels truncate before the
  kebab instead of colliding with it. Works in both agent and customer portals.
- Tickets.vue: drop the now-duplicated local viewActions/handleView/dialog
  helpers and consume them from the composable.
- types: make View.name required (it always exists and the menu relies on it).
- Migrate default view icons from emoji to lucide names (triangle-alert,
  user-check, hourglass, star).
