+++ desk/src/composables/useView.ts
@@ -1,10 +1,19 @@
+import { EditIcon, PinIcon, UnpinIcon } from "@/components/icons";
 import { useAuthStore } from "@/stores/auth";
+import { globalStore } from "@/stores/globalStore";
+import { __ } from "@/translation";
 import { View } from "@/types";
 import { getIcon, isCustomerPortal } from "@/utils";
 import { useDebounceFn } from "@vueuse/core";
-import { call, createListResource, createResource } from "frappe-ui";
-import { computed, ref, watch } from "vue";
-import { useRouter } from "vue-router";
+import {
+  call,
+  createListResource,
+  createResource,
+  FeatherIcon,
+  toast,
+} from "frappe-ui";
+import { computed, h, ref, watch } from "vue";
+import { useRoute, useRouter } from "vue-router";
 
 const debouncedSetValue = useDebounceFn(
   (doctype: string, name: string, fieldname: any, cb?: Function) => {
@@ -37,6 +46,8 @@ export const currentView = ref({
 export function useView(dt: string = null) {
   const auth = useAuthStore();
   const router = useRouter();
+  const route = useRoute();
+  const { $dialog } = globalStore();
   function callGetViews() {
     if (
       (views.filters?.dt === dt && views.data?.length > 0) ||
@@ -180,6 +191,233 @@ export function useView(dt: string = null) {
     };
   }
 
+  // Kebab-menu actions for a single view. Shared between the breadcrumb dropdown
+  // and the sidebar so both expose the same options. `viewDialog` is the caller's
+  // reactive modal state, populated here for the edit/duplicate flows.
+  function viewActions(view: View, viewDialog: any) {
+    const _view = findView(view.name).value;
+
+    const actions: any[] = [
+      {
+        group: __("Default Views"),
+        hideLabel: true,
+        items: [
+          {
+            label: __("Duplicate"),
+            icon: h(FeatherIcon, { name: "copy" }),
+            onClick: () => {
+              viewDialog.view.label = _view.label + " (New)";
+              viewDialog.view.icon = _view.icon;
+              viewDialog.view.name = _view.name;
+              viewDialog.mode = "duplicate";
+              viewDialog.show = true;
+            },
+          },
+        ],
+      },
+    ];
+
+    if (!_view.public || auth.isManager) {
+      if (!_view.public && !_view.is_standard) {
+        actions[0].items.push({
+          label: _view?.pinned ? __("Unpin View") : __("Pin View"),
+          icon: h(_view?.pinned ? UnpinIcon : PinIcon, { class: "h-4 w-4" }),
+          onClick: () => updateView({ name: _view.name, pinned: !_view.pinned }),
+        });
+      }
+      if (_view?.is_standard && auth.isManager) {
+        actions[0].items.push({
+          label: _view?.public
+            ? __("Hide from sidebar")
+            : __("Show in sidebar"),
+          icon: h(FeatherIcon, {
+            name: _view?.public ? "eye-off" : "eye",
+            class: "h-4 w-4",
+          }),
+          onClick: () =>
+            toggleViewVisibility(
+              _view,
+              __("Hide view from sidebar"),
+              __(
+                "{0} view is currently visible in the sidebar. Hiding it will remove it from the sidebar.",
+                [_view.label]
+              )
+            ),
+        });
+      }
+      if (!_view.is_standard) {
+        if (auth.isManager && !isCustomerPortal.value) {
+          actions[0].items.push({
+            label: _view?.public ? __("Make Private") : __("Make Public"),
+            icon: h(FeatherIcon, {
+              name: _view?.public ? "lock" : "unlock",
+              class: "h-4 w-4",
+            }),
+            onClick: () =>
+              toggleViewVisibility(
+                _view,
+                __("Make view private"),
+                __(
+                  "{0} view is currently public. Changing it to private will hide it for all the users.",
+                  [_view.label]
+                )
+              ),
+          });
+        }
+        actions[0].items.push({
+          label: __("Edit"),
+          icon: h(EditIcon, { class: "h-4 w-4" }),
+          onClick: () => {
+            viewDialog.view.label = _view.label;
+            viewDialog.view.icon = _view.icon;
+            viewDialog.view.name = _view.name;
+            viewDialog.mode = "edit";
+            viewDialog.show = true;
+          },
+        });
+        actions.push({
+          group: __("Delete View"),
+          hideLabel: true,
+          items: [
+            {
+              label: __("Delete"),
+              icon: "lucide-trash-2",
+              theme: "red",
+              onClick: () => confirmDeleteView(_view),
+            },
+          ],
+        });
+      }
+    }
+    return actions;
+  }
+
+  function confirmDeleteView(_view: View) {
+    $dialog({
+      title: __("Delete {0}", [_view.label]),
+      message:
+        __("Are you sure you want to delete this view?") +
+        (_view.public
+          ? " " +
+            __("This view is public, and will be removed for all users.")
+          : ""),
+      actions: [
+        {
+          label: __("Confirm"),
+          variant: "solid",
+          iconLeft: "trash-2",
+          theme: "red",
+          onClick({ close }: any) {
+            if (route.query.view === _view.name) {
+              router.push({
+                name: isCustomerPortal.value
+                  ? "TicketsCustomer"
+                  : "TicketsAgent",
+              });
+            }
+            deleteView(_view.name);
+            toast.success(__("View {0}", [__("deleted")]));
+            close();
+          },
+        },
+      ],
+    });
+  }
+
+  function toggleViewVisibility(_view: any, title: string, message: string) {
+    const newView: any = { name: _view.name, public: !_view.public };
+    if (_view.public) {
+      $dialog({
+        title,
+        message,
+        actions: [
+          {
+            label: __("Confirm"),
+            variant: "solid",
+            onClick({ close }: any) {
+              close();
+              updateView(newView);
+            },
+          },
+        ],
+      });
+    } else {
+      updateView(newView);
+    }
+  }
+
+  // Submit handler for the view modal (create/edit/duplicate). `getList` lets the
+  // create flow capture the current list's filters/columns/rows; it is omitted by
+  // callers (e.g. the sidebar) that never create a view.
+  function handleView(
+    viewInfo: any,
+    action: string,
+    viewDialog: any,
+    getList: () => any = () => null
+  ) {
+    let view: View;
+    if (action === "update") {
+      updateView(viewInfo);
+      handleViewSuccess(viewDialog, __("updated"));
+      currentView.value = {
+        label: viewInfo.label,
+        icon: getIcon(viewInfo.icon),
+      };
+      return;
+    } else if (action === "duplicate") {
+      const source = findView(viewDialog.view.name).value;
+      view = {
+        ...source,
+        filters: JSON.stringify(source.filters),
+        columns: JSON.stringify(source.columns),
+        rows: JSON.stringify(source.rows),
+        label: viewInfo.label,
+        icon: viewInfo.icon,
+        public: false,
+        pinned: false,
+      };
+    } else {
+      const list = getList();
+      view = {
+        dt,
+        type: "list",
+        label: viewInfo.label ?? __("List"),
+        icon: viewInfo.icon ?? "",
+        route_name: router.currentRoute.value.name as string,
+        order_by: list?.params.order_by,
+        filters: JSON.stringify(list?.params.filters),
+        columns: JSON.stringify(list?.data.columns),
+        rows: JSON.stringify(list?.data?.rows),
+        is_customer_portal: isCustomerPortal.value,
+      };
+    }
+
+    createView(view, (d) => {
+      currentView.value = {
+        label: d.label || __("List"),
+        icon: getIcon(d.icon),
+      };
+      router.push({
+        name: isCustomerPortal.value ? "TicketsCustomer" : "TicketsAgent",
+        query: { view: d.name },
+      });
+      handleViewSuccess(viewDialog);
+    });
+  }
+
+  function handleViewSuccess(viewDialog: any, msg: string = __("created")) {
+    toast.success(__("View {0}", [msg]));
+    resetViewDialog(viewDialog);
+  }
+
+  function resetViewDialog(viewDialog: any) {
+    viewDialog.show = false;
+    viewDialog.view.label = "";
+    viewDialog.view.icon = "";
+    viewDialog.view.name = "";
+    viewDialog.mode = null;
+  }
+
   watch(
     () => isCustomerPortal.value,
     (newVal) => {
@@ -200,5 +438,8 @@ export function useView(dt: string = null) {
     updateView,
     deleteView,
     createOrUpdateDefaultView,
+    viewActions,
+    handleView,
+    resetViewDialog,
   };
 }
