+++ helpdesk/setup/default_views.py
@@ -8,7 +8,7 @@ default_views = [
         "dt": "HD Ticket",
         "filters": '{"sla":["is","set"],"agreement_status":["in","First Response Due,Resolution Due"],"status_category":"Open","_assign":["LIKE","%@me%"],"creation":["timespan","last 90 days"]}',
         "group_by_field": None,
-        "icon": "⏳",
+        "icon": "triangle-alert",
         "is_customer_portal": 0,
         "is_default": 0,
         "is_standard": 1,
@@ -31,7 +31,7 @@ default_views = [
         "dt": "HD Ticket",
         "filters": '{"__assigned_on":["timespan","last 7 days"],"status_category":"Open","_assign":["LIKE","%@me%"]}',
         "group_by_field": None,
-        "icon": "🆕",
+        "icon": "user-check",
         "is_customer_portal": 0,
         "is_default": 0,
         "is_standard": 1,
@@ -54,7 +54,7 @@ default_views = [
         "dt": "HD Ticket",
         "filters": '{"_assign":["LIKE","%@me%"],"status_category":"Open","last_customer_response":["is","set"],"creation":["timespan","last 90 days"]}',
         "group_by_field": None,
-        "icon": "📋",
+        "icon": "hourglass",
         "is_customer_portal": 0,
         "is_default": 0,
         "is_standard": 1,
@@ -77,7 +77,7 @@ default_views = [
         "dt": "HD Ticket",
         "filters": '{"_assign":["LIKE","%@me%"],"feedback_rating":["is", "set"]}',
         "group_by_field": None,
-        "icon": "⭐",
+        "icon": "star",
         "is_customer_portal": 0,
         "is_default": 0,
         "is_standard": 1,
